+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - CPM-101: Remove twig/extensions dependency (abandoned)
 - CPM-100: replace deprecated `Symfony\Component\Translation\TranslatorInterface\TranslatorInterface` by `Symfony\Contracts\Translation\TranslatorInterface`
+- CPM-100: replace deprecated `Symfony\Component\HttpKernel\Event\GetResponseEvent` by `Symfony\Component\HttpKernel\Event\RequestEvent`
 
 ### Codebase
 
@@ -24,7 +25,6 @@
 - Change constructor of `Akeneo\UserManagement\Bundle\EventListener\LocaleSubscriber` to:
     - remove `Symfony\Component\Translation\TranslatorInterface\TranslatorInterface $translator`
     - add  `Symfony\Contracts\Translation\LocaleAwareInterface\LocaleAwareInterface $localeAware`
-
 
 ### CLI commands